Made AI movement speed relative to hider speed

+++ src/main/java/org/bioshock/engine/core/GameLoop.java
@@ -8,7 +8,7 @@ import org.bioshock.engine.scene.SceneManager;
 import javafx.animation.AnimationTimer;
 
 public final class GameLoop extends AnimationTimer {
-    private static final int LOGICRATE = 60;
+    private static final double LOGICRATE = 60;
 
 	private long prev = 0;
     private long lastUpdate = 0;
@@ -16,7 +16,7 @@ public final class GameLoop extends AnimationTimer {
 	@Override
 	public void handle(long now) {
 		long nanoSDelta = now - prev;
-		double sDelta = nanoSDelta / 1e9;       
+		double sDelta = nanoSDelta / 1e9;
 
         if (now - lastUpdate >= (1/LOGICRATE)*1e9) {
             NetworkManager.tick();
@@ -24,12 +24,12 @@ public final class GameLoop extends AnimationTimer {
             SceneManager.getScene().logicTick(sDelta);
 
             lastUpdate = now;
-        }        
+        }
 
         SceneManager.getScene().renderTick(sDelta);
-        RenderManager.tick();        
+        RenderManager.tick();
         FrameRate.tick(now);
-        
+
 		prev = now;
 	}
 }
